# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para gestión de inventario de camisetas de fútbol y clientes B2B. Proveedor mayorista de camisetas nacionales e internacionales.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
        "/camisetas": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Listar camisetas",
                "description": "Permite filtrar por cliente_id via query param",
                "operationId": "e3fa74d0285600c2a864942784c1d75a",
                "parameters": [
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "description": "Filtrar camisetas por cliente",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de camisetas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Camiseta"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Crear una camiseta",
                "operationId": "36a975dbe7e7989c3ad96430796cfff9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CamisetaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camiseta creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Camiseta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/camisetas/{id}": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Obtener una camiseta con precio final",
                "description": "Si se envía cliente_id y el cliente es Preferencial con precio_oferta definido, precio_final será el precio_oferta. En caso contrario, precio_final será el precio base.",
                "operationId": "54e01d8c44896e1be85402f564c6ad1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "description": "ID del cliente para calcular precio final",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de camiseta con precio_final",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Camiseta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Actualizar una camiseta",
                "operationId": "96d4f09ecf01194b2a6d07609f2dcc75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CamisetaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Camiseta actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Camiseta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Eliminar una camiseta",
                "operationId": "f9e015f046c53518bbe07bb7c8de271c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta eliminada"
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CamisetaInput": {
                "required": [
                    "titulo",
                    "club",
                    "pais",
                    "tipo",
                    "color",
                    "precio",
                    "codigo_producto"
                ],
                "properties": {
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025 – Selección Chilena"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "precio_oferta": {
                        "type": "integer",
                        "example": 38000,
                        "nullable": true
                    },
                    "detalles": {
                        "type": "string",
                        "example": "Edición aniversario 2025",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "tallas": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
            "Camiseta": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025 – Selección Chilena"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "precio_oferta": {
                        "type": "integer",
                        "example": 38000,
                        "nullable": true
                    },
                    "detalles": {
                        "type": "string",
                        "example": "Edición aniversario 2025",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "precio_final": {
                        "type": "integer",
                        "example": 38000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}